lazy-load but keep the loaders colocated to their pages
