+++ apps/react-vite/src/app/routes/app/discussions/discussion.tsx
@@ -1,11 +1,39 @@
+import { QueryClient } from '@tanstack/react-query';
 import { ErrorBoundary } from 'react-error-boundary';
-import { useParams } from 'react-router-dom';
+import { useParams, LoaderFunctionArgs } from 'react-router-dom';
 
 import { ContentLayout } from '@/components/layouts';
 import { Spinner } from '@/components/ui/spinner';
+import { getCommentsQueryOptions } from '@/features/comments/api/get-comments';
 import { Comments } from '@/features/comments/components/comments';
-import { useDiscussion } from '@/features/discussions/api/get-discussion';
+import {
+  useDiscussion,
+  getDiscussionQueryOptions,
+} from '@/features/discussions/api/get-discussion';
 import { DiscussionView } from '@/features/discussions/components/discussion-view';
+
+export const discussionLoader =
+  (queryClient: QueryClient) =>
+  async ({ params }: LoaderFunctionArgs) => {
+    const discussionId = params.discussionId as string;
+
+    const discussionQuery = getDiscussionQueryOptions(discussionId);
+    const commentsQuery = getCommentsQueryOptions(discussionId);
+
+    const promises = [
+      queryClient.getQueryData(discussionQuery.queryKey) ??
+        (await queryClient.fetchQuery(discussionQuery)),
+      queryClient.getQueryData(commentsQuery.queryKey) ??
+        (await queryClient.fetchQuery(commentsQuery)),
+    ] as const;
+
+    const [discussion, comments] = await Promise.all(promises);
+
+    return {
+      discussion,
+      comments,
+    };
+  };
 
 export const DiscussionRoute = () => {
   const params = useParams();
