+++ apps/react-vite/src/app/routes/app/discussions/discussions.tsx
@@ -1,6 +1,18 @@
+import { QueryClient } from '@tanstack/react-query';
+
 import { ContentLayout } from '@/components/layouts';
+import { getDiscussionsQueryOptions } from '@/features/discussions/api/get-discussions';
 import { CreateDiscussion } from '@/features/discussions/components/create-discussion';
 import { DiscussionsList } from '@/features/discussions/components/discussions-list';
+
+export const discussionsLoader = (queryClient: QueryClient) => async () => {
+  const query = getDiscussionsQueryOptions();
+
+  return (
+    queryClient.getQueryData(query.queryKey) ??
+    (await queryClient.fetchQuery(query))
+  );
+};
 
 export const DiscussionsRoute = () => {
   return (
