--- apps/react-vite/src/app/routes/app/discussions/loaders/discussions-loader.ts
@@ -1,12 +1,0 @@
-import { QueryClient } from '@tanstack/react-query';
-
-import { getDiscussionsQueryOptions } from '@/features/discussions/api/get-discussions';
-
-export const discussionsLoader = (queryClient: QueryClient) => async () => {
-  const query = getDiscussionsQueryOptions();
-
-  return (
-    queryClient.getQueryData(query.queryKey) ??
-    (await queryClient.fetchQuery(query))
-  );
-};
