+++ apps/react-vite/src/app/routes/app/users.tsx
@@ -1,6 +1,18 @@
+import { QueryClient } from '@tanstack/react-query';
+
 import { ContentLayout } from '@/components/layouts';
+import { getUsersQueryOptions } from '@/features/users/api/get-users';
 import { UsersList } from '@/features/users/components/users-list';
 import { Authorization, ROLES } from '@/lib/authorization';
+
+export const usersLoader = (queryClient: QueryClient) => async () => {
+  const query = getUsersQueryOptions();
+
+  return (
+    queryClient.getQueryData(query.queryKey) ??
+    (await queryClient.fetchQuery(query))
+  );
+};
 
 export const UsersRoute = () => {
   return (
